Script to spawn enemies in a grid
Currently the script can be placed on an empty game object to spawn a set of enemies in a grid at the scene start and parent them to that game object. The script also passes a waypoint set to the newly spawned enemies.

Currently you can edit the number of columns and rows as well as the space between them, and what waypoint set will be given to the spawned enemies.

I am working on a way to show where the enemies will be spawned in the editor window but it is proving to be difficult.