Editor Extension for the grid spawner tool
Created an editor extension for the grid spawner tool. After inputting the rows, columns, and spacing variables you can press the preview button on the inspector to show where the enemies will spawn. If you change any of those variables you will have to click preview again to update.

Problems that remain: not all of the preview enemies will despawn when you click preview again. This is not an issue on game start but is annoying in the editor.